User Interface Bootstrap Code V0.0.0.0.0.0.0.0.0.1
Okay bois, got vertical tabs for add project, projects, specialists and a spare. The add project one has several horizontal tabs in, the first one has a form in so the operator can enter info but that form probably has questions for second and third tab in it at the moment.

There is an issue though, the vertical tabs are on top of the horizontal tabs not side by side so that is messed up but should be easily fixable.