Add files via upload
Added a table to Problems List with formatting so pending and closed look different. Hover also makes each row slightly darker.

When you click a row it changes tab to a form, where the caller name is automatically filled in. Will have to write an algorithm to populate the other form elements.